remaining clean up including B button fixes
